bump scripts version
