+++ examples/ens-updater/README.md
@@ -5,6 +5,8 @@ Application for updating ENS domain content hash. To send transactions you can u
 You can pass domain and hash to this app by the url. For example: http://ens-updater.loc/#theswarm.eth/2c2d2adb8fd0cba399282fb59f8219e5fbbd67ba06fcf5c8d343f5eb1c8be022
 
 ### Installation
+Project already ready to use without any commands, but if you want make some changes you can follow this instruction:
+
 Inside ens-updater directory run this commands.
 
 
@@ -12,6 +14,7 @@ Inside ens-updater directory run this commands.
 npm install 
 ```
 
+
 ```shell
-bower install 
+browserify js/main.js > bundle.js
 ```
