+++ examples/ens-updater/index.html
@@ -63,7 +63,7 @@
                     </form>
 
                     <p>
-                        <button id="saveDomainHash" class="btn btn-success btn-lg float-right disabled" type="button">
+                        <button id="saveDomainHash" class="btn btn-success btn-lg float-right" type="button">
                             Save
                         </button>
                     </p>
@@ -103,10 +103,7 @@
 <!-- Bootstrap core JavaScript
 ================================================== -->
 <!-- Placed at the end of the document so the pages load faster -->
-<script src="bower_components/web3/dist/web3.js"></script>
-<script src="node_modules/jquery/dist/jquery.min.js"></script>
-<script src="node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
-<script src="js/main.js?1"></script>
-<script src="js/ensutils.js?1"></script>
+<script src="bundle.js"></script>
+
 </body>
 </html>
